Move Plugin to PHPUnit and some other fixes
