+++ library/GeometriaLab/Test/PHPUnit/Controller/AbstractConsoleControllerTestCase.php
@@ -2,8 +2,6 @@
 
 namespace GeometriaLab\Test\PHPUnit\Controller;
 
-use GeometriaLab\Test\Helper\HelperBroker;
-
 /**
  * @method \Zend\Console\Request getRequest()
  * @method \Zend\Console\Response getResponse()
