+++ library/GeometriaLab/Test/PHPUnit/Plugin/Exception/InvalidPluginException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace GeometriaLab\Test\Plugin\Exception;
+namespace GeometriaLab\Test\PHPUnit\Plugin\Exception;
 
 class InvalidPluginException extends \Exception
 {
